server/requirements.txt
fastapi[standard]
pydantic
tortoise-orm